Combine publish and girder dandiset store
Combines the dandiset specific things from the girder and publish state
information into one `dandisets` store module.
