+++ web/src/store/dandiset.js
@@ -1,11 +1,15 @@
-import { publishRest } from '@/rest';
+import { girderRest, publishRest } from '@/rest';
 
 export default {
   namespaced: true,
   state: {
     publishDandiset: null,
+    girderDandiset: null,
   },
   mutations: {
+    setGirderDandiset(state, dandiset) {
+      state.girderDandiset = dandiset;
+    },
     setPublishDandiset(state, dandiset) {
       state.publishDandiset = dandiset;
     },
@@ -14,6 +18,10 @@ export default {
     async fetchPublishDandiset({ commit }, { identifier, version, girderId }) {
       const data = await publishRest.specificVersion(identifier, version, girderId);
       commit('setPublishDandiset', data);
+    },
+    async fetchGirderDandiset({ commit }, { girderId }) {
+      const { data } = await girderRest.get(`folder/${girderId}`);
+      commit('setGirderDandiset', data);
     },
   },
 };
