+++ web/src/store/index.js
@@ -1,7 +1,7 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 
-import publish from '@/store/publish';
+import dandiset from '@/store/dandiset';
 import girder from '@/store/girder';
 import stats from '@/store/stats';
 
@@ -10,7 +10,7 @@ Vue.use(Vuex);
 
 export default new Vuex.Store({
   modules: {
-    publish,
+    dandiset,
     girder,
     stats,
   },
